cleanup: move both converter impls into converter package
