+++ pkg/converter/cli/cli_converter.go
@@ -1,4 +1,4 @@
-package cli_converter
+package cli
 
 import (
 	"encoding/json"
