+++ pkg/converter/cli/cli_converter_test.go
@@ -1,4 +1,4 @@
-package cli_converter
+package cli
 
 import (
 	"testing"
