+++ pkg/converter/cli/llm_tasks.go
@@ -1,4 +1,4 @@
-package cli_converter
+package cli
 
 import (
 	"context"
