+++ pkg/converter/openapi/testdata/openapi_all_invalid_tools.json
