+++ pkg/converter/openapi/testdata/openapi_with_invalid_tools.json
