+++ pkg/converter/openapi/testdata/petstorev3.json
